refactor : model modified
