+++ src/main/java/racing/RacingApplication.java
@@ -10,8 +10,8 @@ public class RacingApplication {
     public static void main(String[] args) {
         String[] carNames = Input.getCarNames();
         int gameCount = Input.getGameCount();
-        RacingGame racingGame = new RacingGame(gameCount);
-        racingGame.gameStart(new Racing(carNames));
+        RacingGame racingGame = new RacingGame();
+        racingGame.gameStart(new Racing(carNames, gameCount));
     }
 
 }
