+++ src/main/java/racing/controller/RacingGame.java
@@ -8,27 +8,12 @@ import java.util.List;
 
 public class RacingGame {
 
-    private static final int END_GAME_COUNT = 0;
-    private int gameCount;
-
-    public RacingGame(int gameCount) {
-        this.gameCount = gameCount;
-    }
-
     public void gameStart(Racing racing) {
         List<Car> carList;
         do {
             carList = racing.game();
             Output.printCarMove(carList);
-        } while (!endGame());
+        } while (!racing.endGame());
         Output.printWinner(racing.winnerCarList());
     }
-
-    public Boolean endGame() {
-        if (this.gameCount > END_GAME_COUNT) {
-            this.gameCount--;
-        }
-        return gameCount == END_GAME_COUNT;
-    }
-
 }
