+++ src/test/java/race/RacingTest.java
@@ -2,11 +2,12 @@ package race;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import race.controller.TestRacingGame;
 import racing.controller.RacingGame;
 import racing.model.Car;
 import racing.model.Racing;
 
+
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -17,10 +18,21 @@ public class RacingTest {
     @Test
     @DisplayName("endGame test")
     void endGameTest() {
-        RacingGame racingGame = new RacingGame(GAME_COUNT);
-        Racing racing = new Racing(new String[]{"test1", "test2", "test3"});
+        RacingGame racingGame = new RacingGame();
+        Racing racing = new Racing(new String[]{"test1", "test2", "test3"}, GAME_COUNT);
         racingGame.gameStart(racing);
-        assertThat(racingGame.endGame()).isEqualTo(true);
+        assertThat(racing.endGame()).isEqualTo(true);
+    }
+
+    @Test
+    @DisplayName("racing test")
+    void racingTest() {
+        Racing racing = new Racing(new String[]{"test1", "test2", "test3"}, GAME_COUNT);
+        List<Car> carList = racing.game();
+        for (Car car : carList) {
+            assertThat(car.totalDistance()).isGreaterThanOrEqualTo(0);
+            assertThat(car.totalDistance()).isLessThanOrEqualTo(5);
+        }
     }
 
 }
